requirements.txt
# Core dependencies
python-dotenv==1.1.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
jsonschema==4.19.2

# Database
sqlite3

# Data processing
pandas==2.1.4
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.9.1
flake8==6.1.0
